requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
astroid==2.5.6
colorama==0.4.4
Django==3.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
future==0.18.2
html5lib==1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pylint==2.8.2
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.67
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
